Move getparameter
